Add cost components for pm-20 WatsonX service (#258)
* initial commit

* added essentials charges

* split up custom resolver

* update unit names for gf

* Add cost components for Standard plan for pm-20

Update golden file

* Fix PR requested changes

* collapse dns cus resolv cc into one func

* Append units to variable names

* Add new model hosting cost components

These cost components are for model hosting with specific
GPUs running per hour. There are 14 new cost components for
the L40S, A100, H100 and H200 GPUs. There are different costs
depending on the number of GPUs running, either in values of
1, 2, 4, or 8.

Instead of creating a seperate cost component function for each,
a generic cost component function called `WMLModelHostingGPUCostComponent`
was created that takes in a unit value from the PostGres db, and maps to
the corresponding title to be used in the golden file. The `WML_ModelHostingHours`
value is created to represent the number of hours for all GPU model hosting/hour
cost components. All other model hosting hours variables are removed.

In the future as more model hosting/hour costs are added, would just need
to update the map with the correct key-value pair.

Updates golden file.

* Update cost component variable names

---------

Co-authored-by: Jack.Cameron@IBM.com <Jack.Cameron@IBM.com>