Add new tests for `container_vpc_cluster and update ibm.go to set `ibm_container_vpc_worker_pool` as child of `ibm_container_vpc_cluster` (#278)
* Add new resources to container_vpc_cluster_test.tf and update ibm.go

This commit updates the container_vpc_cluster_test.tf file to include
a `vpc_cluster` resource that has a `vpc_worker_pool`. This is to verify
the cost esimate of this DA:

https://cloud.ibm.com/catalog/7a4d68b4-cf8b-40cd-a3d1-f49aff526eb3/architecture/deploy-arch-ibm-ocp-vpc-1728a4fd-f561-4cf9-82ef-2b1eeb5da1a8-global

There was some confusion regarding a `Kubernetes Service` being shown along
with the OpenShift cluster on the console. Upon investigation, it appears that the terraform
file for that DA contains both a cluster and worker pool resource. The `Kubernetes Service`
shown on the console represents this worker pool resource. To avoid confusion, an update
is made to `ibm.go` so that the serviceId of `ibm_container_vpc_worker_pool`
is changed from `containers-kubernetes` to `Worker Pool` and is now added
to the childResources property of `ibm_container_vpc_cluster`. This way
worker pools will be displayed as a child of the cluster resource and not
a seperate row.

Updates `container_vpc_cluster_test.golden` file.

* Fix PR comments

Updates golden file.

Add `entitlement` property to worker pool resource so that price matches
what is shown on console.

Update usage.yml so that cluster and pool are both using 730 hours.