Update go.yml to fix GitHub actions go workflow
This commit edits the `go.yml` file for GitHub actions to run the go cli,
format terraform files, and run the IBM unit tests successfully.

Each action is seperated into its own job and runs in parallel.

Fixes formatting for some terraform files because it was failing the job originally.